Extend jetpack features to wings
