+++ src/main/java/com/hbm/items/armor/JetpackBooster.java
@@ -17,7 +17,7 @@ import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.util.Vec3;
 import net.minecraft.world.World;
 
-public class JetpackBooster extends JetpackBase {
+public class JetpackBooster extends JetpackFueledBase {
 
 	public JetpackBooster(FluidType fuel, int maxFuel) {
 		super(fuel, maxFuel);
