+++ src/main/java/com/hbm/items/armor/WingsMurk.java
@@ -1,6 +1,9 @@
 package com.hbm.items.armor;
 
+import java.util.List;
+
 import com.hbm.extprop.HbmPlayerProps;
+import com.hbm.handler.ArmorModHandler;
 import com.hbm.items.ModItems;
 import com.hbm.render.model.ModelArmorWings;
 import com.hbm.util.ArmorUtil;
@@ -10,29 +13,24 @@ import cpw.mods.fml.relauncher.SideOnly;
 import net.minecraft.client.model.ModelBiped;
 import net.minecraft.entity.EntityLivingBase;
 import net.minecraft.entity.player.EntityPlayer;
-import net.minecraft.item.ItemArmor;
 import net.minecraft.item.ItemStack;
 import net.minecraft.util.Vec3;
 import net.minecraft.world.World;
 
-public class WingsMurk extends ItemArmor {
+public class WingsMurk extends JetpackBase {
 
-	public WingsMurk(ArmorMaterial material) {
-		super(material, 0, 1);
+	public WingsMurk() {
+		super();
 	}
-
-	@SideOnly(Side.CLIENT)
-	ModelArmorWings model;
 
 	@Override
 	@SideOnly(Side.CLIENT)
 	public ModelBiped getArmorModel(EntityLivingBase entityLiving, ItemStack itemStack, int armorSlot) {
-		
-		if(model == null) {
-			model = new ModelArmorWings(this == ModItems.wings_murk ? 0 : 1);
+		if(cachedModel == null) {
+			cachedModel = new ModelArmorWings(this == ModItems.wings_murk ? 0 : 1);
 		}
 		
-		return model;
+		return cachedModel;
 	}
 
 	public void onArmorTick(World world, EntityPlayer player, ItemStack stack) {
